+++ examples/resources/morpheus_aws_cloud/import.sh
@@ -1,1 +1,1 @@
-terraform import morpheus_vsphere_cloud.tf_example_aws_cloud 1
+terraform import morpheus_azure_cloud.tf_example_aws_cloud 1